+++ src/Exception/StorageException.java
@@ -1,0 +1,2 @@
+package Exception;public class StorageException {
+}
