+++ src/model/Resume.java
@@ -1,14 +1,12 @@
-package Functional;
+package model;
 
 public class Resume implements Comparable<Resume> {
-    private String uuid;
-
-    public Resume() {
-    }
+    private final String uuid;
 
     public Resume(String uuid) {
         this.uuid = uuid;
     }
+
 
     @Override
     public String toString() {
@@ -19,9 +17,6 @@ public class Resume implements Comparable<Resume> {
         return uuid;
     }
 
-    public void setUuid(java.lang.String uuid) {
-        this.uuid = uuid;
-    }
 
     @Override
     public int compareTo(Resume o) {
